Done refactoring
